LPS-108002 Extract to Keywords component
